WIP: Add shift and mask BLiTTER example
The goal is to make a BLiTTER sprite output.

Vertical coordinate is OK,
for horizontal coordinate BLiTTER can only have delta_x in words,
so for pixel-precise sprites we have to use mask-and-shift.
For example if a sprite pos_x is 1, we use ENDMASK = 0x7fff and
skew = 1, if pos_x is 2, then ENDMASK = 0x3fff and skew = 2, and
so-forth.

Thus sprite output is multi-pass (per right-shifted word) in case
of not word-aligned output position.